refactor standalone to common
