+++ auto-config/core/src/main/yarn/Yarn.java
@@ -1,7 +1,7 @@
 package core.src.main.yarn;
 import java.util.Hashtable;
 
-import core.src.main.standalone.Common;
+import core.src.main.common.Common;
 import utils.UtilsConversion;
 
 public class Yarn {
